Always load konsolepart and port away from KServiceTypeTrader
The current code loads any part of the "TerminalEmulator" type via KServiceTypeTrader.

We are planning to remove KServiceTypeTrader and the replacements don't really cover the use case of "Part of a specific type". KParts::PartLoader only deals with different mimetypes

KonsolePart is the only implementation of a TerminalEmulator part that I am aware of and we might as well load that directly then